roles

+++ app/Http/Controllers/Api/AnomaliaCatalagoController.php
@@ -19,7 +19,7 @@ class AnomaliaCatalagoController extends Controller
      */
     public function index()
     {
-       // $this->authorize('viewAny', AnomaliaCatalogo::class);
+       $this->authorize('viewAny', AnomaliaCatalogo::class);
          
         return AnomaliaCatalogoResource::collection(
             AnomaliaCatalogo::all()
@@ -32,7 +32,7 @@ class AnomaliaCatalagoController extends Controller
      */
     public function store(StoreAnomaliaCatalogoRequest $request)
     {
-       // $this->authorize('create', AnomaliaCatalogo::class);
+       $this->authorize('create', AnomaliaCatalogo::class);
 
         $data = $request->validated();
         $anomalia = AnomaliaCatalogo::create($data);
@@ -59,6 +59,8 @@ class AnomaliaCatalagoController extends Controller
      */
     public function update(UpdateAnomaliaCatalogoRequest $request)
     {
+        $this->authorize('update', AnomaliaCatalogo::class);
+        
         $data = $request->validated();
         $anomalia = AnomaliaCatalogo::find($request["id"]);
         $anomalia->update($data);
@@ -71,6 +73,7 @@ class AnomaliaCatalagoController extends Controller
      */
     public function destroy(Request $request)
     {
+        $this->authorize('delete', AnomaliaCatalogo::class);
         $anomalia = AnomaliaCatalogo::find($request["id"]);
         $anomalia->delete();
     }
